CodeGen.Parse

Documentation

ptr :: Parser ()

ptr2 :: Parser ()

tentypes :: Parser Parsable

ctypes :: Parser Parsable

genericParsers :: forall x. [x] -> (x -> Parsable) -> (x -> Text) -> Parser Parsable

build a parser that will try to find the double-pointer- or pointer- variant first.

typeParser :: (x -> Text) -> x -> Parser x

forLibraries :: (LibType -> Parser x) -> Parser x

parse a library-dependent parser across all of our supported libraries

parsabletypes :: Parser Parsable

api :: Parser ()

semicolon :: Parser ()

functionArg :: Parser Arg

functionArgs :: Parser [Arg]

genericPrefixes :: Parser (LibType, Text)

function :: Parser (Maybe Function)

inlineComment :: Parser ()

comment :: Parser ()

skip over a _single-line_ of block comment -- something which seems standard in the libTH.

parser :: Parser [Maybe Function]

run a parser to find all possible functions, returning one maybe type per-line.

constant :: Parser (Maybe Function)

returns a Maybe Function because we actually don't care about constants when generating FFI code.

skip :: Parser (Maybe Function)

Skip a line because we have failed to find a function