CodeGen.Parse.Cases

Documentation

type2real :: LibType -> TemplateType -> Text

type2hsreal :: TemplateType -> Text

spliced text to use for function names

type2accreal :: LibType -> TemplateType -> Text

checkFunction :: LibType -> TemplateType -> FunctionName -> Bool

Warning a function that doesn't exist will return True by default

TODO: make this safer. (stites): to make this safer I think we need to invert these maps so that we are given function names instead of doing membership checks.