CodeGen.Types.CLI

Documentation

data LibType

All possible libraries that we intend to support (these are all src libraries in ATen). Note that this ordering is used in codegen and must not be changed.

Constructors

ATen

THCUNN

THCS

THC

THNN

THS

TH

prefix :: LibType -> Bool -> Text

describe' :: LibType -> String

Short descriptions of each library we intend to support.

supported :: LibType -> Bool

Whether or not we currently support code generation for the library

supportedLibraries :: [LibType]

outDir :: LibType -> FilePath

Where generated code will be placed.

outModule :: LibType -> Text

The prefix of the output module name

srcDir :: LibType -> CodeGenType -> FilePath

Where the source files are located, relative to the root of the hasktorch project.

data CodeGenType

Type of code to generate

Constructors

GenericFiles
generic/ files which are used in C for type-generic code

ConcreteFiles
concrete supporting files. These include utility functions and random generators.

generatable :: CodeGenType -> Bool

Whether or not we currently support generating this type of code (ie: I (@stites) am not sure about the managed files).

data TemplateType

Constructors

GenByte

GenChar

GenDouble

GenFloat

GenHalf

GenInt

GenLong

GenShort

GenNothing

generatedTypes :: LibType -> CodeGenType -> [TemplateType]